+++ questioner/urls.py
@@ -43,16 +43,13 @@ schema_view = get_schema_view(
 )
 urlpatterns = [
     path("", views.Index.as_view(), name="welcome"),
-    path("auth/login/", views.Login.as_view(), name="login"),
-    path("auth/signup/", views.SignUp.as_view(), name="signup"),
+    path("accounts/", include('accounts.urls')),
     path("meetups/", include("meetup.urls")),
     path("meetups/", include("question.urls")),
     path("tags/", meetup_views.TagList.as_view(), name="tags"),
     path("tags/<int:tag_id>", meetup_views.ATag.as_view(), name="tag"),
     path("admin/", admin.site.urls),
-    path('accounts/', include('rest_framework.urls', namespace='rest_framework')),
-    path("accounts/signup/", views.SignUp.as_view(), name="signup2"),
-    path("accounts/profile/", views.profile.as_view(), name="signup2"),
+    path('session/', include('rest_framework.urls', namespace='rest_framework')),
     path(
         "docs/",
         schema_view.with_ui("swagger", cache_timeout=0),
